requirements.txt
streamlit==1.49.1
pillow==11.3.0
numpy==2.3.2
torch==2.8.0+cpu
torchvision==0.23.0+cpu